requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
gunicorn>=20.1.0
aiohttp>=3.8.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
jinja2>=3.0.0
python-multipart>=0.0.5
requests>=2.26.0
pydantic>=1.8.0
typing-extensions>=4.0.0
asyncio>=3.4.3
aiofiles>=0.8.0
python-json-logger>=2.0.0
prometheus-client>=0.12.0  # For metrics
sentry-sdk>=1.5.0  # For error tracking